backend/requirements.txt
# Core Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Supabase
supabase>=2.10.0
httpx>=0.28.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0

# Database (for migrations)
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.14.0
sqlalchemy>=2.0.0
